+ Added new ChessBoardSampler
This sampler uses either an "A" or "B" sampler, depending on the UV coordinate
in in a chess board pattern.

This works the same as the ChessBoard material, but is simpler, since it is only
a sampler.

If the same material model is used for the whole object, this sampler could be
used e.g. to provide different textures, while using the same lighting equation.