* Rework Cone and Cylinder intersection
The existing intersection implementations for Cone and Cylinder geometries had
many subtle (and a few not-so subtle) errors.

Intersections from inside the geometries were not properly modelled, and many
edge cases were just not right.

After much work, there's now a new, more extensively tested implementation,
which seems to bring us much closer to a stable, correct implementation.

The cylinder is currently implemented as a degenerate cone, so there is a lot of
room for optimization.

Also make sure the `quadratic2` helper function does not try to sort results,
since this leads to incorrect results.