* Major rework of material and sampler ui code
All materials now implement SceneObject and Interactive, and we draw their ui
using these traits.

Macros are used to implement SceneObject, much like it is done for Geometry.

Samplers and Materials no longer create a CollapsibleHeader and/or a Grid. This
is now instead done in `frontend::gui`, and the Interactive objects just fill in
their own information.